changed all the code so instead of writing while reading, we store the info in a string
Thanks to this change, we are able to, if there is a fail (fail to allocate memory),
return a -1 and not print anything. This way, the printf functionality is closer to
the original functionality, and it makes sense that if we return -1 we don't print.
The drawback is that the function is more complicated, and we need to constantly
manage memory.